pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tydline-core"
version = "0.1.0"
description = "Tydline - early warning container intelligence backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.8.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "groq>=0.5.0",
    "pydantic-ai>=0.0.14",
    "mem0ai>=1.0.0",
    "langfuse>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx[cli]>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
